refactor: form edit and add collection
